Improve error messages
Using Error might be bad for Fetcher if hyper errors are common. The performance
penalty might require us to switch to #[derive(Fail)] or an Error struct and
ErrorKind enum approach.

Also, it might be annoying to have to convert Error to &Fail each time we want
to print it. But this way let's us use only one function.