Only update last_fetched on success
Previously, we updated last_fetched as we were creating the future. But, if the
request failed, then last_fetched would be incorrect. This could result in lost
updates. Now, last_fetched is only updated after the request succeeds.

There is a tiny inaccuracy: for the ease of chaining futures, last_fetched is
updated in a chained future to the actual first future to touch the response.
However, this should be fine, as the futures should not have a giant gap between
when they run.