Allow trivial_regex, remove TODO comment about Cow
We can't actually let the caller substitute their own string if the final result
is Cow::Borrowed, because that only tells us if the last regex in the chain did
any replacements or not. If any earlier regex did a replacement, the Cow::Owned
would be "covered up" by the final regex.

So, we need to perform at least one unnecessary clone no matter what. A small
consolation is that this is fairly cheap, since usernames and titles (what this
function is only applied to) tend to be short.

We also disable Clippy's lint, because despite the regexes being trivial,
Regex::replace_all does not copy the string if no matches are found, whereas
str::replace always copies the whole string.