Remove the threshold heuristic for deleted threads
The threshold heuristic was copied from Asagi, but the anchor index heuristic is
better. With it and a threshold of 7, the only time the threshold heuristic
activates is if the last 30 threads are modified or bumped off, and the 31st to
last thread is deleted.

That's unlikely to happen unless the poll interval is high--in which case the
heuristic is useless because so much time has passed that it's plausible the
"deleted" thread really was bumped off. So, the threshold heuristic is unneeded.