Use Last-Modified with FetchThread
This is a partial reversion of 0ef03539. We now send Last-Modified for
threads because of the delay between BoardPoller checking threads.json
and ThreadUpdater actually fetching the thread.

If many updates happen quickly, by the time ThreadUpdater fetches the
thread, it will see newer data than BoardPoller. At the next poll,
BoardPoller will catch up to ThreadUpdater, and send it a redundant
modification message. By checking Last-Modified, we ensure that this
message does not cause an unnecessary database write.