Purge Last-Modified hashmap every day
An day should be long enough that few active threads are purged. On slow
boards, many threads will be purged, but this is fine because those
threads weren't active anyways. It's unlikely timing issues will cause
redundant updates with them.

It's also okay to use such a large interval because each DateTime<Utc>
is only 12 bytes, so hardly any memory is leaked.