Separate thread and thread list rate limiting
When we add support for polling the archives, we must ensure that the
threads.json and archive.json for each board are polled in fairly quick
succession. This ensures that we don't miss any threads archived between
the polls of archive.json and threads.json.

For example, with the old model, after polling archive.json, we add all
of its threads to the thread queue. threads.json gets shoved to the very
end of this queue, which delays its request, causing us to miss archived
threads until the queue clears up.

By adding a separate queue for thread lists, we ensure that these
requests are not delayed. This, in turn, allows us to poll the archives
just once, as opposed to creating an ArchivePoller.