Fix media downloading
The previous method had two flaws, which are now fixed.

1. It used "total = 1" to check if a media was new or not. However, when
inserts with the same image run concurrently, the total can get set to 2
before either NEW_MEDIA_QUERY runs. Thus, a full media is never
downloaded. We check against the filename (which doesn't change once it
is set) to prevent this.

2. It assumed that inserts inserted rows consecutively. However, this is
not the case when many inserts are running at once. Thus, each selection
of rows between LAST_INSERT_ID() and LAST_INSERT_ID() + ROW_COUNT() - 1
was returning many rows from other threads, causing duplicate downloads.
Now, we only select rows from the current thread and posts we are
inserting. We also check the database for existing posts
beforehand (NEXT_NUM_QUERY) and ignore them, in case we are updating.

The conclusion from all this is to always assume that the database has
changed between queries, and write code to defend against that.