Adjust rate limiting parameters
Statistics for images (full and spoiler) per thread from a single scrape:
    Mean:    57.94
    Std dev: 53.23
    Min:      1.75
    25th:    18.04
    Median:  34.15
    75th:   102.06
    Max:    185.07

This skew-right distribution suggests that the current max_interval
ratio of 2 images to 1 thread is too high. On start, the image queue
will get bogged down. Thus, the thread max_interval has been decreased
to 200. This is still high, but lessens the pressure. Setting it to 10
would ensure that images are processed at the same rate as threads, but
is too low to handle updates after initialization.