Crash system when media runtime panics
Because the future to send a message to the media rate limiter is run on
a separate Runtime, if a media fetch future crashes, the whole system
will keep running and spew out "receiver closed" messages. Now we
explicitly panic if the sender is closed so that the system crashes when
the media runtime crashes.