Replace html5ever with regex and Pest
Switching to regular expressions and Pest
=========================================

I initially chose to use html5ever instead of regular expressions (as
Asagi does) because regexes (from the regex crate), even with lazy
matching, can't parse nested tags correctly. For example, this HTML:

<span class="mu-r"><span class="mu-b"></span></span>

is transformed into this BBCode by regex:

[qstcolor=red][b][/qstcolor][/b]

The regex fails to correctly match start and ending <span> tags, and
produces mismatched BBCode tags as a result.

However, using html5ever just to solve this problem was a mistake. We
end up paying performance and complexity penalties for features that we
don't use, such as: (note that mod/sticky posts are always an exception)

- Full HTML spec support: 4chan only uses about 11 HTML elements.
- Malformed HTML parsing: The HTML in posts is generated from BBCode and
  other markup, so we can assume it is always valid.
- Document tree manipulation: Transforming HTML to BBCode preserves the
  original structure of the comment.
- etc.

As it turns out, we can use regex for most of the parsing, and Pest, a
PEG parser generator, to handle nested tags. This simplifies the code
and, according to unscientific benchmarks, improves parsing speed by
6.5x. Other optimizations include:

- Only parsing and replacing when there are tags or entities in the
  input. This is done by checking for `<` or `&`. This means there is no
  performance penalty for cleaning or unescaping plaintext posts.
- Avoiding regex captures. For simple, unique tags, we just replace the
  start and end tags with BBCode. For complex tags (spans, [code],
  [banned], etc.), we use the Pest parser.
- Looping over `Matches` instead of calling `replace_all` for each tag.

Other changes
=============

- Removed "just-in-case" support for syntax that hasn't been seen in the
  wild (<b> [banned], <sub>, <sub>, "#fff" shorthand [fortune] colors,
  etc). These elements can be added back if they are needed.
- The [fortune] tag now removes the newlines and <b> tag from inside it.
- EXIF line breaks are removed by regex instead of by string trimming.
  This means blankposts will no longer be altered.
- Attributes and text are not purposely escaped as required by the HTML
  spec. This is in line with Asagi. Again, because FoolFuuka escapes
  everything, this isn't a problem.

Caveat
======

Whereas regex ignores malformed HTML, if the Pest parser is fed
mismatched or malformed tags, it will error and stop parsing. Thus, in
the unlikely case the parser is given input with valid elements hidden
among malformed ones, the valid ones will not be parsed into BBCode.
But again, this is unlikely to happen.