Create media fetch futures on-demand
The media sender now accepts FetchMedia messages, which are lazily
flattened and mapped to futures. This saves channel memory, because we
don't have to queue tons of futures, and saves sender memory, because we
only clone media_rl_sender once for each thread. Previously, the sender
was cloned for each media file.