Check media path, use more async fs
The standard library currently barely supports file permissions, so we
create a test file to see if we can access the media directory.

Continuing the flip-flopping of creating the "tmp" directory versus
saving a syscall (5b54f71 and 657167c), the temp directory is now
created before every temp file is opened. Even though this is
"wasteful", the cost of checking that a directory exists is small enough
that it should be fine. This method is also more robust, and will not
cause media downloads to fail just because a directory is missing.

The `exists` call in fetch_media remains synchronous because its purpose
is to exit early if the media already exists. Checking it in the future
would always incur the cost of creating and queuing the future and
negate its purpose. This check should be fast enough that task
throughput will not be affected.