Increase rate limiting and base backoff delay
4chan API rules state that no more than one request per second should be
made. Though the actual limit is higher (see issue 57 in 4chan-API),
defaults should try to follow the rules.

The defaults are set so that a maximum of three requests will be made
per second. This is higher than the limit. But:

- The number of threads and images to fetch is unpredictable. If we set
  the limits too low, we may end up blocking requests. So, we set the
  limits above the recommendation. Assuming that the user isn't scraping
  many boards, these limits should be fine.
- Thread lists are not fetched often. So, though the higher value
  could cause the request rate to exceed 2 req/s, it usually won't.

The base delay of the backoff was increased to 8 seconds. Retrying too
early will probably result in another error, which is wasteful.