Don't error on empty thread list/archive
An empty thread is invalid because a thread must always have at least
one post (the OP). But, an empty thread list is not invalid. During a
raid, for example, it could be possible for the mods to delete every
thread on the board. An empty archive is also not invalid. If a board is
very slow, days or weeks could go by without a single new thread,
allowing the archive to empty out.

(In practice, it's unlikely that either of these scenarios will happen.
An empty thread list or archive may be more likely to be a database
glitch. But, we trust the API. So, we take what we get at face value.)