Don't match nested tags in fortune, adjust comments
Trying to use `value*` in fortune means that `text` must not match `<b>`
and `</b>`. This runs into problems, because we end up not matching
`<br>` as well. Using "b>" | "b " works, but is hacky because the space
is not the only valid whitespace after a tag name.

To save the annoyance of having to write out all the possible whitespace
characters, I only match text in fortune. It's unlikely that tags will
be nested in a fortune tag anyway.

I also removed the note about not using regex. By applying its logic, we
should match all tags with Pest, as any tag could be reused. What if <a>
links had classes? What if <b> banned tags came back?

To avoid unnecessary logic, I assume that <pre> and <strong> are the
only tags that might be reused in the future. This keeps the SIMPLE_TAGS
regex simple and lets us strip links easily. There's always the small
risk of losing data, but I think the cleaner code outweighs the risk.