Revise doc comments
The comments for the config structs were not useful. It's fine to leave
the structs uncommented because this isn't a library.

Describing the HTML to BBCode conversion as "partial" is confusing. It
makes it seem like we don't convert all of the 4chan HTML into BBCode,
which we do. What I meant is that we only convert a subset of all HTML
into BBCode. If a mod inserts a <h1> tag into a post, we won't convert
it. However, that's a given, so we don't need to mention it.